requirements.txt
streamlit
langchain
langchain-groq
langchain-community
fpdf
requests
PyPDF2
python-docx
